indexer/pyproject.toml
[project]
name = "indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=2.2.0",
    "requests>=2.28.0",
    "pyarrow>=21.0.0",
    "pandas>=2.3.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "torchaudio>=2.6.0",
    "num2words>=0.5.14",
    "timm>=1.0.19",
    "accelerate>=1.10.1",
    "transformers>=4.44.0",
    "flash-attn",
    "openai>=1.102.0",
    "python-dotenv>=1.0.1",
    "peft>=0.15.2",
    "pillow>=11.0.0",
    "websockets>=15.0.1",
    "opencv-python>=4.12.0.88",
]

# PyTorch CUDA wheels index
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# Optional CPU index if needed
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu",    marker = "sys_platform == 'darwin'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu",    marker = "sys_platform == 'darwin'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu",    marker = "sys_platform == 'darwin'" },
]

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]

[tool.uv.extra-build-variables]
flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }
